Exit non-zero for compile failures
A potential point of failure in this bash script is related to the curl request which fetches the binary from `DOWNLOAD_URL`, perhaps due to a transient network error.

This commit adds bash strict mode to ensure that the heroku build process will fail, in the event any bash error occurs by returning a non-zero status code.